Memory efficiency changes. Mainly restructuring structs and reducing size of numbers required to store

+++ src/cn_access.h
@@ -23,10 +23,11 @@
 #define _cn_access_h
 
 #include <stdio.h>
+#include <stdint.h>
 
-int cn_access_get_copy_number_for_location(char *file_loc,char *chr,int pos,int is_normal);
+int8_t cn_access_get_copy_number_for_location(char *file_loc,char *chr,uint32_t pos,uint8_t is_normal);
 void clear_copy_number_store();
 void cn_access_set_max_cn(int max_copy_number);
-int cn_access_get_mean_cn_for_range(char *file_loc,char *chr,int start,int stop,int is_normal);
+int8_t cn_access_get_mean_cn_for_range(char *file_loc,char *chr,uint32_t start,uint32_t stop,uint8_t is_normal);
 
 #endif
